Move navigation translations to core/utils
