+++ public/app/core/utils/navBarItem-translations.ts
